fix(e2e): getByLabel('Nom') bloquait 30s sur le formulaire d'item
Le run CI de #238 a révélé que item-management.e2e.test.ts et
quantity-update.e2e.test.ts échouaient tous les deux sur le même
timeout : getByLabel('Nom', { exact: true }) ne matchait jamais.

Cause : le <label htmlFor="item-label"> contient "Nom" + un span
aria-hidden="true" pour l'astérisque de champ requis. Le texte réel
utilisé par getByLabel est "Nom *" (aria-hidden n'est pas pris en
compte pour ce matching), donc exact: true sur 'Nom' ne matche jamais
— contrairement à l'accessible name du textbox associé (calculé
différemment, lui exclut bien le "*").

Remplacé par des sélecteurs par id (#item-label, #item-quantity,
#item-minimum-stock), le pattern déjà validé pour le formulaire de
stock. Factorisé dans tests/e2e-frontend/helpers/item-actions.ts
(addItem) pour éviter de dupliquer cette logique dans les deux tests.
